+++ literary_system/critic/__init__.py
@@ -43,3 +43,5 @@ from literary_system.critic.next_episode_bench import (run_next_episode_bench, N
 from literary_system.critic.distribution_guard import (distribution_guard, GuardResult, compute_stats, apply_guard_to_reward, NORMAL_BANDS)  # V784 M3
 
 from literary_system.critic.self_eval_pipeline import (SelfEvalPipeline, SelfEvalReport)  # V785
+
+from literary_system.critic.distribution_guard import (distribution_guard_features, CORPUS_FEATURE_BANDS)  # V787